Moved src/test files to src/main
