+++ src/main/java/com/mongodb/diffutil/DiffUtilRunner.java
@@ -5,8 +5,6 @@ import java.util.Set;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
-import com.mongodb.diffutil.DiffSummary;
-import com.mongodb.diffutil.DiffUtil;
 import com.mongodb.model.Namespace;
 import com.mongodb.mongopush.config.MongoPushConfiguration;
