+++ src/main/java/com/mongodb/mongopush/model/MongoPushTestEvent.java
@@ -1,4 +1,4 @@
-package com.mongodb.mongopush.events;
+package com.mongodb.mongopush.model;
 
 public enum MongoPushTestEvent {
